deployment/requirements.txt
click
python-dotenv
pyyaml
paramiko
ansible
rpyc